+++ lib/dalli/protocol/value_marshaller.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require 'forwardable'
+require 'debug'
 
 module Dalli
   module Protocol