+++ lib/dalli/protocol/value_serializer.rb
@@ -10,6 +10,7 @@ module Dalli
     ##
     class ValueSerializer
       DEFAULTS = {
+        raw: false,
         serializer: Marshal
       }.freeze
 
@@ -27,7 +28,7 @@ module Dalli
       end
 
       def store(value, req_options, bitflags)
-        do_serialize = !(req_options && req_options[:raw])
+        do_serialize = serialize_value?(req_options)
         store_value = do_serialize ? serialize_value(value) : value.to_s
         bitflags |= FLAG_SERIALIZED if do_serialize
         [store_value, bitflags]
@@ -60,6 +61,19 @@ module Dalli
         exc = Dalli::MarshalError.new(e.message)
         exc.set_backtrace e.backtrace
         raise exc
+      end
+
+      private
+
+      # Defaults to client-level option, unless a request-level override is provided
+      def serialize_value?(req_options)
+        return !raw_by_default? unless req_options && !req_options[:raw].nil?
+
+        !req_options[:raw]
+      end
+
+      def raw_by_default?
+        @serialization_options[:raw]
       end
     end
   end